下班

+++ WhereToBuy-3/WhereToBuy/Classes/Other/Tools.h
@@ -37,6 +37,6 @@
 //获取Tmp目录
 +(void)dirTmp;
 
-- (CGFloat) folderSizeAtPath:(NSString*) folderPath;
+- (NSInteger) folderSizeAtPath:(NSString*) folderPath;
 
 @end
